Changes 4 Til
Secondarys are optional so not implementing them to win condition triggers. Players will naturaly kill the officer when clearing the sawmill and if the ZU truck gets destroyed then they just wont have succeeded with that secondary.

Per Woody's advice I've culled a squad from each QRF wave, he says this will be ok for the number of players.