TVT (60-79) VDV Hello From The Sky V1 | ВДВ, с неба привет!
Soviet VDV Company with Attached Mi-24s are conducting an air assault to secure the Bysovice Airfield from Canadian control, having already bypassed most of the forward emplaced defences. Canadians are just now aware of the threat and are preparing defences while QRFs are inbound, but unknown when or if they'll arrive.